Refactor parseLocationForWhitelistedModule
