+++ packages/typescript-resolver-files/src/utils/parseLocationForWhitelistedModule.spec.ts
